remove last of the facades
