--- src/facade/mod.rs
@@ -1,1 +1,0 @@
-pub mod metadata_facade;
